# test(vue): Switch to using vitest