--- a/packages/vue/jest.config.js
+++ b/packages/vue/jest.config.js
@@ -1,6 +1,0 @@
-const baseConfig = require('../../jest/jest.config.js');
-
-module.exports = {
-  ...baseConfig,
-  testEnvironment: 'jsdom',
-};