… packages/vue/test/vendor/components.test.ts