Before: `Time: 3.935 s`

After: `Duration  1.55s (transform 961ms, setup 0ms, collect 3.56s, tests 127ms, environment 1.73s, prepare 367ms)`

If we remove the jsdom based tests it'll probably get even faster.

ref https://github.com/getsentry/sentry-javascript/issues/11084

earlier attempt: https://github.com/getsentry/sentry-javascript/pull/11071

I'll try to do a couple of these a week when I'm blocked on some of my other tasks.